Fix CONTRIB-6917.

+++ classes/observer.php
@@ -41,15 +41,18 @@ class format_grid_observer {
      * @param \core\event\course_content_deleted $event
      */
     public static function course_content_deleted(\core\event\course_content_deleted $event) {
-        global $DB;
-        /* Delete any images associated with the course.
-           Done this way so will work if the course has
-           been a grid format course in the past even if
-           it is not now. */
-        $courseformat = format_grid::get_instance($event->objectid);
-        $courseformat->delete_images();
-        unset($courseformat);  // Destruct.
+        if (class_exists('format_grid', false)) {
+            // If class format_grid was never loaded, this is definitely not a course in 'Grid' format.
+            global $DB;
+            /* Delete any images associated with the course.
+               Done this way so will work if the course has
+               been a grid format course in the past even if
+               it is not now. */
+            $courseformat = format_grid::get_instance($event->objectid);
+            $courseformat->delete_images();
+            unset($courseformat);  // Destruct.
 
-        $DB->delete_records("format_grid_summary", array("courseid" => $event->objectid));
+            $DB->delete_records("format_grid_summary", array("courseid" => $event->objectid));
+        }
     }
 }
